requirement.txt
tenacity
openai
backoff
tiktoken
python-dotenv